% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperFunctions.R
\name{analizeEvolution}
\alias{analizeEvolution}
\title{Analysis of repeated measures of the same patient over time
Analyze the evolution of a patient over time. This function returns the Placido irregularity indices per time step
and two temporal plots.}
\usage{
analizeEvolution(data, fileExtension = NULL)
}
\arguments{
\item{data}{Either 1) the path of a folder that contains corneal topography files, 
as exported by Placido disks corneal topographers, or 2) a list containing properly formatted data
(loaded from a file using the function \link[rPACI]{readCornealTopography}, 
simulated using \link[rPACI]{simulateData}, or by other ways, as long as it meets the dataset requirements).}

\item{fileExtension}{If data is a path, specify the file extension of the corneal topography files 
in the folder. It assumes all files with the given extension are corneal topography files of a single patient.}
}
\description{
Analysis of repeated measures of the same patient over time
Analyze the evolution of a patient over time. This function returns the Placido irregularity indices per time step
and two temporal plots.
}
\details{
If the data are loaded from a folder, it will be assumed that the temporal arrangement is the alphabetical order of the files' name. 
Therefore, it is advised to use a proper file name, for instance using this date format: 'YYYY-MM-DD.txt'.
Moreover, the folder should contain data measures of one patient only since the function 
will read all the files (with the given extension) of the specified folder.
On the other hand, if the data are stored in a list, it will be assumed that the temporal order corresponds with the index of the dataset in the list.
}
\examples{
# EXAMPLE 1
# Simulate the patient's measures over time
 dataT1 = simulateData(rings = 12, maximumMireDisplacement = 0.15, mireDisplacementAngle = 10)
 dataT2 = simulateData(rings = 12, maximumMireDisplacement = 0.15, mireDisplacementAngle = 45)
 dataT3 = simulateData(rings = 12, maximumMireDisplacement = 0.2, mireDisplacementAngle = 50)
 
# Create a list containing the data
data = list(
 dataT1 = dataT1,
 dataT2 = dataT2,
 dataT3 = dataT3
)

# Analyze the data over time
analizeEvolution(data)

# EXAMPLE 2
# Specify a folder path to analyze a patient's evolution over time
analizeEvolution(system.file("extdata", package="rPACI"), fileExtension = 'txt')
}
