% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeData.R
\name{writerPACI}
\alias{writerPACI}
\title{Write a dataset in the format used by rPACI}
\usage{
writerPACI(df, filename, sep = ",")
}
\arguments{
\item{df}{A \code{data.frame} containing the corneal topography points, with columns:
\tabular{lll}{
  \code{x}   \tab\tab The X Cartesian coordinates of the points\cr
  \code{y}   \tab\tab The Y Cartesian coordinates of the points\cr
  \code{ring index}  \tab\tab Number or index of the ring to which each point belongs\cr
}}

\item{filename}{Character string naming a file (including extension).}

\item{sep}{The field separator string. By default, ','.}
}
\description{
Write a corneal topography dataset in the format used by rPACI.
This format consists of an optional header of any length (its size is automatically detected) and afterwards, 
three columns (x and y coordinates of each point and its ring index) and a row per data point.
This format is returned by the functions \link[rPACI]{readCSO} and \link[rPACI]{simulateData}.
}
